# Easy-ToDoList-main/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.LoginReq:
    properties:
      account:
        type: string
      password:
        type: string
    type: object
  dto.RegisterReq:
    properties:
      account:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.ToDoAddReq:
    properties:
      description:
        type: string
      end_time:
        type: string
      title:
        type: string
    type: object
  dto.ToDoUpdateReq:
    properties:
      description:
        type: string
      end_time:
        type: string
      status:
        type: integer
      title:
        type: string
    type: object
  dto.UpdateUserInfoReq:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  vo.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  vo.LoginResponse:
    properties:
      message:
        example: 登录成功
        type: string
      token:
        example: some.jwt.token
        type: string
      user:
        $ref: '#/definitions/vo.UserVO'
    type: object
  vo.RegisterResponse:
    properties:
      message:
        example: 注册成功
        type: string
      user:
        $ref: '#/definitions/vo.UserVO'
    type: object
  vo.SuccessResponse:
    properties:
      message:
        example: success
        type: string
    type: object
  vo.ToDoVO:
    properties:
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
  vo.UpdateUserResponse:
    properties:
      message:
        example: 更新成功
        type: string
      user:
        $ref: '#/definitions/vo.UserVO'
    type: object
  vo.UserVO:
    properties:
      account:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a todo list.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ToDo List API
  version: "1.0"
paths:
  /todos/add:
    post:
      consumes:
      - application/json
      description: Add a new todo item to the list
      parameters:
      - description: Todo item to add
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.ToDoAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Create a new todo item
      tags:
      - todos
  /todos/delete/{id}:
    delete:
      description: Delete a specific todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Delete a todo item by ID
      tags:
      - todos
  /todos/get/{id}:
    get:
      description: Get a specific todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ToDoVO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Get a todo item by ID
      tags:
      - todos
  /todos/list:
    get:
      description: Get a list of all todo items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vo.ToDoVO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: List all todo items for a user
      tags:
      - todos
  /todos/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a specific todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo item to update
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.ToDoUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Update a todo item by ID
      tags:
      - todos
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: User login
      tags:
      - users
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Register a new user
      tags:
      - users
  /user/update:
    put:
      consumes:
      - application/json
      description: Update the authenticated user's information
      parameters:
      - description: User info to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: Update user information
      tags:
      - users
swagger: "2.0"